docker_project/import_service_process/requirements.txt
psycopg2-binary==2.9.3
sqlalchemy==1.4.37
kafka-python==2.0.2
pandas==1.4.2
sqlalchemy-redshift==0.8.9